chart

+++ pppppp/Model/ChartsWeightItem.swift
@@ -2,6 +2,6 @@ import Foundation
 
 struct ChartsWeightItem: Identifiable {
     var date: String
-    var weight: Float
+    var weight: Double
     var id = UUID()
 }
